requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
psycopg2-binary==2.9.9
httpx==0.25.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
